yagoutpay-backend/requirements.txt
Flask==2.3.3
pycryptodome==3.19.0